Move subclass definitions to .cpp files and remove unused includes
Address PR review comments:
- Move one-liner test subclass declarations from fixture headers into
  their corresponding .cpp files since each header is only included
  by a single source file.
- Remove unused includes (<cstdint>, <type_traits>, <hip/hip_runtime.h>)
  from all three fixture headers.
